Defect #11 (Partial): Director values need polish
Simplify director system to two states of relax and build up: in relax the director doesn't have actions so there are no positive goals. The next stage of the rework will involve breaking down the goals - they are too general for the director to determine the advantage of different actions right now so we will break them down into different skills, different needs and the madness goal. The values of these items will then translate to insistency values (the higher the skill, the higher the insistency) (the higher the need, the lower the insistency) (the higher the madness chance, the lower the insistency)